// admin/src/store/home/type.ts
export interface IHomeState {
  userListData: any[];
  userListTotal: number;
  movieListData: any[];
  movieListTotal: number;
  cinemaListData: any[];
  cinemaListTotal: number;
  hallListData: any[];
  hallListTotal: number;
  planListData: any[];
  planListTotal: number;
  advertiseListData: any[];
  advertiseListTotal: number;
  commentListData: any[];
  commentListTotal: number;
  orderListData: any[];
  orderListTotal: number;
}